% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_generic.R
\name{as.classCoMiRe}
\alias{as.classCoMiRe}
\title{classCoMiRe class constructor}
\usage{
as.classCoMiRe(call = NULL, out = NULL, z = NULL, z.val = NULL, f0 = NULL, f1 = NULL, 
nrep, nb, bin = FALSE, univariate = TRUE)
}
\arguments{
\item{call}{a formula for \code{comire.gibbs}.}

\item{out}{an output of \code{comire.gibbs}.}

\item{z}{optional numeric vector or matrix for the confounding covariates.}

\item{z.val}{optional numeric vector containing a fixed value of interest for each of the confounding covariates to be used for the plots. Default value is \code{mean(z)} for numeric covariates or the mode for factorial covariates.}

\item{f0, f1}{optional matrices containing simulated values of the mixture densities at low and high dose exposure; default values are simulated with \code{comire.gibbs}. It is possible to change these for differente fixed values of \code{z}: see \code{predict_new_z} function.}

\item{nrep}{integer giving the total number of iterations used in \code{comire.gibbs}.}

\item{nb}{integer giving the number of burn-in iterations used in \code{comire.gibbs}.}

\item{bin}{logical. It is \code{TRUE} if y is drawn for a binomial distribution.}

\item{univariate}{logical. It is \code{TRUE} if the model is univariate.}
}
\description{
A constructor for the \code{classCoMiRe} class. The class \code{classCoMiRe} is a named list containing
the output of the posterior estimation of CoMiRe model implemented in \code{comire.gibbs}
}
\author{
Antonio Canale, Arianna Falcioni
}
