% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_generic.R
\name{bmd.plot}
\alias{bmd.plot}
\title{Benchmark dose plot}
\usage{
bmd.plot(bmd.data)
}
\arguments{
\item{bmd.data}{output of \code{BMD} function.}
}
\description{
Posterior mean (continuous lines) and pointwise credible bands (shaded areas) for the benchmark dose in function of the increase in risk.
}
\author{
Antonio Canale
}
