% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_generic.R
\name{BMD}
\alias{BMD}
\title{Benchmark dose}
\usage{
BMD(level, risk, x, alpha=0.05)
}
\arguments{
\item{level}{dose level of interest.}

\item{risk}{\code{summary.risk$mcmc.risk} from the output of \code{add.risk} function.}

\item{x}{numeric vector for the covariate relative to the dose of exposure used in \code{comire.gibbs}.}

\item{alpha}{level of the credible bands.}
}
\value{
A dataframe containing as variables:
\itemize{
\item{\code{q}}{ the dose level of interest.}
\item{\code{BMD}}{ the benchmark dose.}
\item{\code{low}}{ lower credible limit.}
\item{\code{upp}}{ upper credible limit.}
\item{\code{BMDL}}{ a more conservative benchmark dose.}
}
}
\description{
Benchmark dose associated to a particular risk
}
\author{
Antonio Canale
}
