\name{CoMiRe-package}
\alias{CoMiRe-package}
\alias{CoMiRe}
\docType{package}
\title{
\packageTitle{CoMiRe}
}
\description{
\packageDescription{CoMiRe}
}
\details{
The \code{CoMiRe} package implements the convex mixture regresion approach of Canale, Durante, and Dunson (2018) and some extensions to deal with binary response variables or to account for the presence of continuous and categorical confunders. Estimation is conducted via Gibbs sampler. Posterior plots for inference and goodness-of-fit tests are also available.}
\author{
\packageAuthor{CoMiRe}
Maintainer: \packageMaintainer{CoMiRe}
}
\references{
Canale, A., Durante, D., and Dunson, D. (2018), Convex Mixture Regression for Quantitative Risk Assessment, Biometrics, 74, 1331-1340
}
\keyword{ package }
