% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comire.gibbs.binary.R
\name{comire.gibbs.binary}
\alias{comire.gibbs.binary}
\title{Gibbs sampler for CoMiRe model with binary response}
\usage{
comire.gibbs.binary(
  y,
  x,
  mcmc,
  prior,
  seed,
  max.x = max(x),
  min.x = min(x),
  verbose = TRUE
)
}
\arguments{
\item{y}{numeric vector for the response.}

\item{x}{numeric vector for the covariate relative to the dose of exposure.}

\item{mcmc}{a list giving the MCMC parameters. It must include the following integers: \code{nb} giving the number of burn-in iterations, \code{nrep} giving the total number of iterations, \code{thin} giving the thinning interval, \code{ndisplay} giving the multiple of iterations to be displayed on screen while the algorithm is running (a message will be printed every \code{ndisplay} iterations).}

\item{prior}{a list containing the values of the hyperparameters. 
It must include the following values: 
\itemize{
\item \code{eta}, numeric vector of size \code{J} for the Dirichlet prior on the beta basis weights, 
\item \code{a.pi0} and \code{b.pi0}, the prior parameters of the prior beta distribution for \eqn{\pi_0},
\item \code{J}, parameter controlling the number of elements of the Ispline basis.
}}

\item{seed}{seed for random initialization.}

\item{max.x}{maximum value allowed for \code{x}.}

\item{min.x}{minimum value allowed for \code{x}.}

\item{verbose}{logical, if \code{TRUE} a message on the status of the MCMC algorithm is printed to the console. Default is \code{TRUE}.}
}
\description{
Posterior inference via Gibbs sampler for CoMiRe model with binary response
}
