% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comire.gibbs.continuous.confunders.R
\name{comire.gibbs.continuous.confunders}
\alias{comire.gibbs.continuous.confunders}
\title{Gibbs sampler for CoMiRe model with continuous response and more than one confounder variable.}
\usage{
comire.gibbs.continuous.confunders(
  y,
  x,
  z,
  grid = NULL,
  mcmc,
  prior,
  state = NULL,
  seed = 5,
  max.x = ceiling(max(x)),
  z.val = NULL,
  verbose = TRUE
)
}
\arguments{
\item{y}{numeric vector for the response.}

\item{x}{numeric vector for the covariate relative to the dose of exposure.}

\item{z}{numeric vector for the confunders; a vector if there is only one confounder or a matrix for two or more confunders.}

\item{grid}{a list giving the parameters for plotting the posterior mean density and the posterior mean \eqn{\beta(x)} over finite grids.}

\item{mcmc}{a list giving the MCMC parameters. It must include the following integers: \code{nb} giving the number of burn-in iterations, \code{nrep} giving the total number of iterations, \code{thin} giving the thinning interval, \code{ndisplay} giving the multiple of iterations to be displayed on screen while the algorithm is running (a message will be printed every \code{ndisplay} iterations).}

\item{prior}{a list containing the values of the hyperparameters. 
It must include the following values: 
\itemize{
\item \code{mu.theta}, the prior mean \eqn{\mu_\theta} for each location parameter \eqn{\theta_{0h}}{\theta_0h} and \eqn{\theta_1}, 
\item \code{k.theta}, the prior variance \eqn{k_\theta} for each location paramter \eqn{\theta_{0h}}{\theta_0h} and \eqn{\theta_1}, 
\item \code{mu.gamma} (if \code{p} confounding covariates are included in the model) a \code{p}-dimentional vector of prior means \eqn{\mu_\gamma}{\mu_gamma} of the parameters \eqn{\gamma} corresponding to the confounders,
\item \code{k.gamma}, the prior variance \eqn{k_\gamma}{k_gamma} for parameter corresponding to the confounding covariate (if \code{p=1}) or \code{sigma.gamma} (if \code{p>1}), that is the covariance matrix \eqn{\Sigma_\gamma}{\Sigma_gamma} for the parameters corresponding to the \code{p} confounding covariates; this must be a symmetric positive definite matrix.
\item \code{eta}, numeric vector of size \code{J} for the Dirichlet prior on the beta basis weights, 
\item \code{alpha}, prior for the mixture weights,
\item \code{a} and \code{b}, prior scale and shape parameter for the gamma distribution of each precision parameter, 
\item \code{J}, parameter controlling the number of elements of the I-spline basis,
\item \code{H}, total number of components in the mixture at \eqn{x_0}.
}}

\item{state}{if \code{family="continuous"}, a list giving the current value of the parameters. This list is used if the current analysis is the continuation of a previous analysis or if we want to start the MCMC algorithm from some particular value of the parameters.}

\item{seed}{seed for random initialization.}

\item{max.x}{maximum value allowed for \code{x}.}

\item{z.val}{optional numeric vector containing a fixed value of interest for each of the confounding covariates to be used for the plots. Default value is \code{mean(z)} for numeric covariates or the mode for factorial covariates.}

\item{verbose}{logical, if \code{TRUE} a message on the status of the MCMC algorithm is printed to the console. Default is \code{TRUE}.}
}
\description{
Posterior inference via Gibbs sampler for CoMiRe model with continuous response and more than one confounder variable.
}
