% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comire.internal.R
\name{comire.internal}
\alias{comire.internal}
\alias{pssq_gaussian}
\title{Internal functions of the package}
\usage{
pssq_gaussian(index, data, cluster, locations)
}
\arguments{
\item{index}{index}

\item{data}{data}

\item{cluster}{cluster}

\item{locations}{locations}
}
\description{
Internal functions of the package
}
\keyword{internal}
\keyword{pssq_gaussian}
