% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{labelling_b_C}
\alias{labelling_b_C}
\title{Labelling b}
\usage{
labelling_b_C(w, phi, f0i, f1i)
}
\arguments{
\item{w}{w}

\item{phi}{phi}

\item{f0i}{f0i}

\item{f1i}{f1i}
}
\description{
Labelling b
}
