% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comire.internal.R
\name{labelling_b_uni}
\alias{labelling_b_uni}
\title{labelling_b_uni}
\usage{
labelling_b_uni(i, w, phi, f0i, f1i)
}
\arguments{
\item{i}{i}

\item{w}{w}

\item{phi}{phi}

\item{f0i}{f0i}

\item{f1i}{f1i}
}
\description{
labelling_b_uni
}
