% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{labelling_bb_C}
\alias{labelling_bb_C}
\title{Labelling bb}
\usage{
labelling_bb_C(w, phi, P0, P1)
}
\arguments{
\item{w}{w}

\item{phi}{phi}

\item{P0}{P0}

\item{P1}{P1}
}
\description{
Labelling bb
}
