% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{labelling_c_C}
\alias{labelling_c_C}
\title{Labelling c}
\usage{
labelling_c_C(y, logpi, mu, tau)
}
\arguments{
\item{y}{y}

\item{logpi}{logpi}

\item{mu}{mu}

\item{tau}{tau}
}
\description{
Labelling c
}
