% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comire.internal.R
\name{mixdensity_uni}
\alias{mixdensity_uni}
\title{mixdensity_uni}
\usage{
mixdensity_uni(i, y, z, nu, theta, tau, ga)
}
\arguments{
\item{i}{i}

\item{y}{y}

\item{z}{z}

\item{nu}{nu}

\item{theta}{theta}

\item{tau}{tau}

\item{ga}{ga}
}
\description{
mixdensity_uni
}
