% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_generic.R
\name{predict_new_z}
\alias{predict_new_z}
\title{comire.gibbs for different fixed values of z}
\usage{
predict_new_z(fit, y, z.val)
}
\arguments{
\item{fit}{the output of \code{comire.gibbs} opportunely trasformed in \code{classCoMiRe} class}

\item{y}{numeric vector for the response used in \code{comire.gibbs}}

\item{z.val}{optional numeric vector containing a fixed value of interest for each of the confounding covariates to be used for the plots. Default value is \code{mean(z)} for numeric covariates or the mode for factorial covariates.}
}
\value{
An object of class \code{classCoMiRe}.
}
\description{
This function computes the predicted values of the density al low 
dose \code{f_0} and of the density at high dose \code{f_{inf}}, for fixed values 
of the confounders \code{z}.
}
\author{
Antonio Canale, Arianna Falcioni
}
