% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_generic.R
\name{print.classCoMiRe}
\alias{print.classCoMiRe}
\title{CoMiRe print}
\usage{
\method{print}{classCoMiRe}(x, ...)
}
\arguments{
\item{x}{an object of class \code{classCoMiRe};}

\item{...}{additional arguments.}
}
\description{
The \code{print.classCoMiRe} method prints the type of a \code{classCoMiRe} object.
}
\author{
Antonio Canale, Arianna Falcioni
}
