% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comire.internal.R
\name{psdp_multi}
\alias{psdp_multi}
\title{psdp_multi}
\usage{
psdp_multi(index, y, z, cluster)
}
\arguments{
\item{index}{index}

\item{y}{y}

\item{z}{z}

\item{cluster}{cluster}
}
\description{
psdp_multi
}
