% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comire.internal.R
\name{psdp_uni}
\alias{psdp_uni}
\title{psdp_uni (posterior double product)}
\usage{
psdp_uni(index, y, z, cluster)
}
\arguments{
\item{index}{index}

\item{y}{y}

\item{z}{z}

\item{cluster}{cluster}
}
\description{
psdp_uni (posterior double product)
}
