% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comire.internal.R
\name{pssq_multi}
\alias{pssq_multi}
\title{pssq_multi}
\usage{
pssq_multi(y, times, z, gamma, theta)
}
\arguments{
\item{y}{y}

\item{times}{times}

\item{z}{z}

\item{gamma}{gamma}

\item{theta}{theta}
}
\description{
pssq_multi
}
