% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comire.internal.R
\name{pssq_uni}
\alias{pssq_uni}
\title{pssq_uni}
\usage{
pssq_uni(index, y, z, cluster, theta, gamma)
}
\arguments{
\item{index}{index}

\item{y}{y}

\item{z}{z}

\item{cluster}{cluster}

\item{theta}{theta}

\item{gamma}{gamma}
}
\description{
pssq_uni
}
