% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_generic.R
\name{riskplot}
\alias{riskplot}
\title{Additional risk function plot}
\usage{
riskplot(risk.data, xlab = NULL, x = NULL, ylim=c(0,1), xlim=c(0, max(x)))
}
\arguments{
\item{risk.data}{output of \code{add.risk} function.}

\item{xlab}{the title of the x axis.}

\item{x}{numeric vector for the covariate relative to the dose of exposure used in \code{comire.gibbs}.}

\item{xlim, ylim}{numeric vectors of length 2, giving the x and y coordinates ranges for the plot.}
}
\description{
Posterior mean (continuous lines) and pointwise credible bands (shaded areas) for \eqn{Ra(x, a)}.
}
\author{
Antonio Canale
}
