% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_generic.R
\name{summary.classCoMiRe}
\alias{summary.classCoMiRe}
\title{CoMiRe summary}
\usage{
\method{summary}{classCoMiRe}(object, ...)
}
\arguments{
\item{object}{an object of class \code{classCoMiRe};}

\item{...}{additional arguments}
}
\description{
The \code{summary.classCoMiRe} method provides summary information on \code{classCoMiRe} objects.
}
\author{
Antonio Canale Arianna Falcioni
}
