\name{CoMiRe-package}
\alias{CoMiRe-package}
\alias{CoMiRe}
\docType{package}
\title{
\packageTitle{CoMiRe}
}
\description{
\packageDescription{CoMiRe}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{CoMiRe}
\packageIndices{CoMiRe}
}
\author{
\packageAuthor{CoMiRe}

Maintainer: \packageMaintainer{CoMiRe}
}
\references{
Canale, A., Durante, D., and Dunson, D. (2017), Convex Mixture Regression for Quantitative Risk Assessment }
\keyword{ package }