% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/comire.internal.R
\name{comire.internal}
\alias{comire.internal}
\alias{.labelling_b_C}
\alias{.labelling_bb_C}
\alias{.labelling_c_C}
\alias{.mixdensity_C}
\alias{.pssq_gaussian}
\alias{.labelling_b_uni}
\alias{.labelling_c_uni}
\alias{.mixdensity_uni}
\alias{.pssq_uni}
\alias{.psdp_uni}
\alias{.labelling_b_multi}
\alias{.labelling_c_multi}
\alias{.mixdensity_multi}
\alias{.pssq_multi}
\alias{.psdp_multi}
\title{Internal Functions of CoMiRe package}
\usage{
.labelling_b_C(w, phi, f0i, f1i)

.labelling_bb_C(w, phi, P0, P1)

.labelling_c_C(y, logpi, mu, tau)

.mixdensity_C(y, pi, mu, tau)

.pssq_gaussian(index, data, cluster, locations)

.labelling_b_uni(i, w, phi, f0i, f1i)

.labelling_c_uni(i, y, z, nu, theta, tau, ga)

.mixdensity_uni(i, y, z, nu, theta, tau, ga)

.pssq_uni(index, y, z, cluster, theta, gamma)

.psdp_uni(index, y, z, cluster)

.labelling_b_multi(i, w, phi, f0i, f1i)

.labelling_c_multi(i, y, z, nu, theta, tau, ga)

.mixdensity_multi(i, y, z, nu, theta, tau, ga)

.pssq_multi(y, times, z, gamma, theta)

.psdp_multi(index, y, z, cluster)
}
\description{
Internal Functions of CoMiRe package
}
\keyword{internal}
